Minor bug fix

+++ Machine Learning Project/ReLU.cpp
@@ -2,7 +2,7 @@
 
 Eigen::MatrixXd ReLU::forward(Eigen::MatrixXd panIn)
 {
-    this->panOut = (panIn.array() >= 0).matrix();
+    this->panOut = (panIn.array() >= 0).cast<double>().matrix();
     Eigen::MatrixXd Out = (this->panOut.array() * panIn.array()).matrix();
     return Out;
 }
